refactor(snapshot): isolate backend capability fixtures
