+++ apple/runner/AgentDeviceRunner/AgentDeviceRunnerUITests/RunnerTests+SnapshotCapturePlan.swift
@@ -42,65 +42,6 @@ struct SnapshotCustomActionCoverage: Codable {
   let blocked: Bool
 }
 
-enum SnapshotBackendKind: String, CaseIterable {
-  case recursiveTree = "tree"
-  case querySweep = "queries"
-  case privateAX = "private-ax"
-
-  var isForceable: Bool {
-    switch self {
-    case .recursiveTree, .privateAX:
-      return true
-    case .querySweep:
-      return false
-    }
-  }
-
-  var usesXCTestAccessibilityChannel: Bool {
-    switch self {
-    case .recursiveTree, .querySweep:
-      return true
-    case .privateAX:
-      return false
-    }
-  }
-
-  /// The raw projection is the acquired tree, so only a backend that enumerates a hierarchy can
-  /// serve it. The query sweep answers an interactive element query: it has no hierarchy to
-  /// return, and planning it for `--raw` is exactly how a raw request gets answered with regular
-  /// membership (#1797 D4).
-  var supportsRawProjection: Bool {
-    switch self {
-    case .recursiveTree, .privateAX:
-      return true
-    case .querySweep:
-      return false
-    }
-  }
-
-  var isAvailableOnCurrentPlatform: Bool {
-    #if os(iOS) && targetEnvironment(simulator)
-      return isAvailable(on: .simulator)
-    #else
-      return isAvailable(on: .physicalDevice)
-    #endif
-  }
-
-  func isAvailable(on environment: SnapshotBackendEnvironment) -> Bool {
-    switch self {
-    case .recursiveTree, .querySweep:
-      return true
-    case .privateAX:
-      return environment == .simulator
-    }
-  }
-}
-
-enum SnapshotBackendEnvironment {
-  case simulator
-  case physicalDevice
-}
-
 enum SnapshotXCTestChannelPlanState: Equatable {
   case normal
   case deferredToIndependentBackend
@@ -137,22 +78,6 @@ struct SnapshotBackendCapture {
   var qualityPayload: DataPayload? = nil
 }
 
-private struct SnapshotBackendParityFixture: Decodable {
-  struct Availability: Decodable {
-    let simulator: Bool
-    let physicalDevice: Bool
-  }
-
-  struct Backend: Decodable {
-    let name: String
-    let forceable: Bool
-    let supportsRawProjection: Bool
-    let availability: Availability
-  }
-
-  let backends: [Backend]
-}
-
 extension RunnerTests {
   static let sparseRecoveryTruncatedNodeThreshold = 8
   /// Umbrella wall-clock budget for one capture plan. Individual backends bound themselves,
@@ -173,22 +98,6 @@ extension RunnerTests {
   /// projection drops out of the raw plan by construction, and a new one joins it by declaring the
   /// trait instead of by someone remembering this line.
   static let rawDiagnosticPlan: [SnapshotBackendKind] = regularVisiblePlan.filter(\.supportsRawProjection)
-
-  private func loadSnapshotBackendParityFixture() throws -> SnapshotBackendParityFixture {
-    let fixtureURL = URL(fileURLWithPath: #filePath)
-      .deletingLastPathComponent() // AgentDeviceRunnerUITests
-      .deletingLastPathComponent() // AgentDeviceRunner
-      .deletingLastPathComponent() // runner
-      .deletingLastPathComponent() // apple
-      .deletingLastPathComponent() // repo root
-      .appendingPathComponent("contracts")
-      .appendingPathComponent("fixtures")
-      .appendingPathComponent("ios-snapshot-backends.json")
-    return try JSONDecoder().decode(
-      SnapshotBackendParityFixture.self,
-      from: Data(contentsOf: fixtureURL)
-    )
-  }
 
   // MARK: XCTest accessibility channel penalty (cross-attempt memory, #1105/#1156)
   //
@@ -951,36 +860,6 @@ extension RunnerTests {
     XCTAssertTrue(Self.rawDiagnosticPlan.allSatisfy(\.supportsRawProjection))
     // Tree-first error propagation is the raw plan's other contract (ADR 0004).
     XCTAssertEqual(Self.rawDiagnosticPlan.first, .recursiveTree)
-  }
-
-  /// The JSON table is the cross-runtime declaration used by the TypeScript capability registry
-  /// and this runner. A backend case, forceability branch, raw projection claim, or availability
-  /// change that is not classified in both implementations fails before an iOS smoke can drift.
-  func testSnapshotBackendDeclarationsMatchCapabilityFixture() throws {
-    let fixture = try loadSnapshotBackendParityFixture()
-    XCTAssertEqual(
-      fixture.backends.map(\.name),
-      SnapshotBackendKind.allCases.map(\.rawValue)
-    )
-
-    for expected in fixture.backends {
-      guard let backend = SnapshotBackendKind(rawValue: expected.name) else {
-        XCTFail("fixture contains an unknown snapshot backend: \(expected.name)")
-        continue
-      }
-      XCTAssertEqual(backend.isForceable, expected.forceable, expected.name)
-      XCTAssertEqual(backend.supportsRawProjection, expected.supportsRawProjection, expected.name)
-      XCTAssertEqual(
-        backend.isAvailable(on: .simulator),
-        expected.availability.simulator,
-        "simulator availability: \(expected.name)"
-      )
-      XCTAssertEqual(
-        backend.isAvailable(on: .physicalDevice),
-        expected.availability.physicalDevice,
-        "physical-device availability: \(expected.name)"
-      )
-    }
   }
 
   /// A projection mismatch is a runner bug, not an accessibility failure: it must not take the
