+++ test/integration/smoke-ios-snapshot-backend-conformance.test.ts
@@ -1,45 +1,20 @@
 import assert from 'node:assert/strict';
 import test from 'node:test';
 
-import type { SnapshotQualityVerdict } from '@agent-device/kernel/snapshot';
 import {
   assertSnapshotBackendConformance,
   createSnapshotBackendConformanceTransport,
   loadSnapshotBackendConformanceFixture,
 } from './ios-simulator-e2e/snapshot-backend-conformance.ts';
+import {
+  buildSnapshotBackendConformanceBase,
+  type SnapshotBackendConformanceInput,
+} from './ios-simulator-e2e/snapshot-backend-conformance-fixtures.ts';
 
 const fixture = loadSnapshotBackendConformanceFixture();
 
 test('snapshot backend conformance checks each backend contract independently', () => {
-  const base = {
-    truncated: false,
-    snapshotQuality: { state: 'healthy', backend: 'tree' } satisfies SnapshotQualityVerdict,
-    nodes: [
-      {
-        index: 0,
-        ref: 'e1',
-        identifier: 'field-name',
-        label: 'Full name',
-        type: 'TextField',
-        value: 'Ada Lovelace',
-        enabled: true,
-        hittable: false,
-        rect: { x: 0, y: 0, width: 100, height: 20 },
-      },
-      {
-        index: 1,
-        ref: 'e2',
-        identifier: 'field-email',
-        label: 'Email',
-        type: 'TextField',
-        value: 'ada@example.test',
-        enabled: true,
-        hittable: false,
-        rect: { x: 0, y: 20, width: 100, height: 20 },
-      },
-      { index: 2, ref: 'e3', type: 'ScrollView' },
-    ],
-  };
+  const base = buildSnapshotBackendConformanceBase();
 
   assert.doesNotThrow(() => assertSnapshotBackendConformance(base, 'tree', fixture));
   assert.throws(
@@ -54,39 +29,14 @@ test('snapshot backend conformance checks each backend contract independently', 
 });
 
 test('snapshot backend conformance rejects every promised control invariant', () => {
-  type SnapshotInput = Parameters<typeof assertSnapshotBackendConformance>[0];
-  const base: SnapshotInput = {
-    truncated: false,
-    snapshotQuality: { state: 'healthy', backend: 'tree' } satisfies SnapshotQualityVerdict,
-    nodes: [
-      {
-        index: 0,
-        ref: 'e1',
-        identifier: 'field-name',
-        label: 'Full name',
-        type: 'TextField',
-        value: 'Ada Lovelace',
-        enabled: true,
-        hittable: false,
-        rect: { x: 0, y: 0, width: 100, height: 20 },
-      },
-      {
-        index: 1,
-        ref: 'e2',
-        identifier: 'field-email',
-        label: 'Email',
-        type: 'TextField',
-        value: 'ada@example.test',
-        enabled: true,
-        hittable: false,
-        rect: { x: 0, y: 20, width: 100, height: 20 },
-      },
-      { index: 2, ref: 'e3', type: 'ScrollView' },
-    ],
-  };
+  const base = buildSnapshotBackendConformanceBase();
   const updateControl = (identifier: string, update: Record<string, unknown>) =>
     base.nodes.map((node) => (node.identifier === identifier ? { ...node, ...update } : node));
-  const expectFailure = (name: string, snapshot: SnapshotInput, message: RegExp) =>
+  const expectFailure = (
+    name: string,
+    snapshot: SnapshotBackendConformanceInput,
+    message: RegExp,
+  ) =>
     assert.throws(() => assertSnapshotBackendConformance(snapshot, 'tree', fixture), message, name);
 
   expectFailure(
